fix panic when file/dir is deleted
sometimes when a file/dir which is being watched is
deleted a panic will occur. This appears to be because
the is an assumption that if an error returns true for
`IsNotExist()` that it should be castable to `os.PathError`,
but this evidently is not always the case, causing a panic.